Move the core of the primitives addon into the core.
This is the first step of #1003, enabling future rendering unification changes.

On OpenGL, this is feature neutral, as the implementation is mostly copy-paste.

On Direct3D this is more involved, as we remove the second shader system and use core's shaders. I believe this should have similar hardware support to the old system, but is far simpler.
