+++ addons/native_dialog/gtk_dialog.c
@@ -1,6 +1,7 @@
 #include <gtk/gtk.h>
 #include <gdk/gdkx.h>
 
+#define ALLEGRO_INTERNAL_UNSTABLE
 #include "allegro5/allegro.h"
 #include "allegro5/allegro_native_dialog.h"
 #include "allegro5/internal/aintern_native_dialog.h"
