+++ addons/primitives/allegro5/allegro_primitives.h
@@ -2,6 +2,7 @@
 #define __al_included_allegro5_allegro_primitives_h
 
 #include <allegro5/allegro.h>
+#include <allegro5/internal/aintern_primitives_types.h>
 
 #if (defined ALLEGRO_MINGW32) || (defined ALLEGRO_MSVC) || (defined ALLEGRO_BCC32)
 #ifndef ALLEGRO_STATICLINK
@@ -31,57 +32,6 @@ extern "C"
 #endif
 
 
-/* Enum: ALLEGRO_PRIM_TYPE
- */
-typedef enum ALLEGRO_PRIM_TYPE
-{
-  ALLEGRO_PRIM_LINE_LIST,
-  ALLEGRO_PRIM_LINE_STRIP,
-  ALLEGRO_PRIM_LINE_LOOP,
-  ALLEGRO_PRIM_TRIANGLE_LIST,
-  ALLEGRO_PRIM_TRIANGLE_STRIP,
-  ALLEGRO_PRIM_TRIANGLE_FAN,
-  ALLEGRO_PRIM_POINT_LIST,
-  ALLEGRO_PRIM_NUM_TYPES
-} ALLEGRO_PRIM_TYPE;
-
-enum
-{
-   ALLEGRO_PRIM_MAX_USER_ATTR = _ALLEGRO_PRIM_MAX_USER_ATTR
-};
-
-/* Enum: ALLEGRO_PRIM_ATTR
- */
-typedef enum ALLEGRO_PRIM_ATTR
-{
-   ALLEGRO_PRIM_POSITION = 1,
-   ALLEGRO_PRIM_COLOR_ATTR,
-   ALLEGRO_PRIM_TEX_COORD,
-   ALLEGRO_PRIM_TEX_COORD_PIXEL,
-   ALLEGRO_PRIM_USER_ATTR,
-   ALLEGRO_PRIM_ATTR_NUM = ALLEGRO_PRIM_USER_ATTR + ALLEGRO_PRIM_MAX_USER_ATTR
-} ALLEGRO_PRIM_ATTR;
-
-/* Enum: ALLEGRO_PRIM_STORAGE
- */
-typedef enum ALLEGRO_PRIM_STORAGE
-{
-   ALLEGRO_PRIM_FLOAT_2,
-   ALLEGRO_PRIM_FLOAT_3,
-   ALLEGRO_PRIM_SHORT_2,
-   ALLEGRO_PRIM_FLOAT_1,
-   ALLEGRO_PRIM_FLOAT_4,
-   ALLEGRO_PRIM_UBYTE_4,
-   ALLEGRO_PRIM_SHORT_4,
-   ALLEGRO_PRIM_NORMALIZED_UBYTE_4,
-   ALLEGRO_PRIM_NORMALIZED_SHORT_2,
-   ALLEGRO_PRIM_NORMALIZED_SHORT_4,
-   ALLEGRO_PRIM_NORMALIZED_USHORT_2,
-   ALLEGRO_PRIM_NORMALIZED_USHORT_4,
-   ALLEGRO_PRIM_HALF_FLOAT_2,
-   ALLEGRO_PRIM_HALF_FLOAT_4
-} ALLEGRO_PRIM_STORAGE;
-
 /* Enum: ALLEGRO_LINE_JOIN
  */
 typedef enum ALLEGRO_LINE_JOIN
@@ -104,60 +54,11 @@ typedef enum ALLEGRO_LINE_CAP
    ALLEGRO_LINE_CAP_CLOSED
 } ALLEGRO_LINE_CAP;
 
-/* Enum: ALLEGRO_PRIM_BUFFER_FLAGS
- */
-typedef enum ALLEGRO_PRIM_BUFFER_FLAGS
-{
-   ALLEGRO_PRIM_BUFFER_STREAM       = 0x01,
-   ALLEGRO_PRIM_BUFFER_STATIC       = 0x02,
-   ALLEGRO_PRIM_BUFFER_DYNAMIC      = 0x04,
-   ALLEGRO_PRIM_BUFFER_READWRITE    = 0x08
-} ALLEGRO_PRIM_BUFFER_FLAGS;
-
-/* Enum: ALLEGRO_VERTEX_CACHE_SIZE
- */
-#define ALLEGRO_VERTEX_CACHE_SIZE 256
 
 /* Enum: ALLEGRO_PRIM_QUALITY
  */
 #define ALLEGRO_PRIM_QUALITY 10
 
-/* Type: ALLEGRO_VERTEX_ELEMENT
- */
-typedef struct ALLEGRO_VERTEX_ELEMENT ALLEGRO_VERTEX_ELEMENT;
-
-struct ALLEGRO_VERTEX_ELEMENT {
-   int attribute;
-   int storage;
-   int offset;
-};
-
-/* Type: ALLEGRO_VERTEX_DECL
- */
-typedef struct ALLEGRO_VERTEX_DECL ALLEGRO_VERTEX_DECL;
-
-/* Duplicated in allegro5/internal/aintern_tri_soft.h */
-#ifndef _ALLEGRO_VERTEX_DEFINED
-#define _ALLEGRO_VERTEX_DEFINED
-
-/* Type: ALLEGRO_VERTEX
- */
-typedef struct ALLEGRO_VERTEX ALLEGRO_VERTEX;
-
-struct ALLEGRO_VERTEX {
-  float x, y, z;
-  float u, v;
-  ALLEGRO_COLOR color;
-};
-#endif
-
-/* Type: ALLEGRO_VERTEX_BUFFER
- */
-typedef struct ALLEGRO_VERTEX_BUFFER ALLEGRO_VERTEX_BUFFER;
-
-/* Type: ALLEGRO_INDEX_BUFFER
- */
-typedef struct ALLEGRO_INDEX_BUFFER ALLEGRO_INDEX_BUFFER;
 
 ALLEGRO_PRIM_FUNC(uint32_t, al_get_allegro_primitives_version, (void));
