+++ addons/primitives/polygon.c
@@ -16,10 +16,11 @@
  *      See readme.txt for copyright information.
  */
 
+#include <math.h>
+
 #include "allegro5/allegro.h"
 #include "allegro5/allegro_primitives.h"
-#include "allegro5/internal/aintern_prim.h"
-#include <math.h>
+#include "allegro5/internal/aintern_prim_addon.h"
 
 #ifdef ALLEGRO_MSVC
    #define hypotf(x, y) _hypotf((x), (y))
