+++ addons/primitives/polyline.c
@@ -16,12 +16,13 @@
  *      See readme.txt for copyright information.
  */
 
+#include <float.h>
+#include <math.h>
+
 #include "allegro5/allegro.h"
 #include "allegro5/allegro_primitives.h"
 #include "allegro5/internal/aintern_list.h"
-#include "allegro5/internal/aintern_prim.h"
-#include <float.h>
-#include <math.h>
+#include "allegro5/internal/aintern_prim_addon.h"
 
 
 /*
