+++ src/memblit.c
@@ -20,6 +20,7 @@
 #include "allegro5/internal/aintern_convert.h"
 #include "allegro5/internal/aintern_memblit.h"
 #include "allegro5/internal/aintern_transform.h"
+#include "allegro5/internal/aintern_primitives.h"
 #include "allegro5/internal/aintern_tri_soft.h"
 #include <math.h>
