+++ src/point_soft.c
@@ -19,12 +19,10 @@
 
 #define _AL_NO_BLEND_INLINE_FUNC
 
-#define ALLEGRO_INTERNAL_UNSTABLE
 #include "allegro5/allegro.h"
-#include "allegro5/allegro_primitives.h"
 #include "allegro5/internal/aintern_bitmap.h"
 #include "allegro5/internal/aintern_blend.h"
-#include "allegro5/internal/aintern_prim.h"
+#include "allegro5/internal/aintern_primitives.h"
 #include "allegro5/internal/aintern_prim_soft.h"
 #include <math.h>
