+++ src/shader_source.inc
@@ -188,8 +188,6 @@ static const char *default_glsl_minimal_pixel_source_gl3 =
 #endif /* ALLEGRO_CFG_SHADER_GLSL */
 
 
-#ifdef ALLEGRO_CFG_SHADER_HLSL
-
 static const char *default_hlsl_vertex_source =
    "struct VS_INPUT\n"
    "{\n"
@@ -222,6 +220,9 @@ static const char *default_hlsl_vertex_source =
    "   Output.Position = mul(Input.Position, " ALLEGRO_SHADER_VAR_PROJVIEW_MATRIX ");\n"
    "   return Output;\n"
    "}\n";
+
+
+#ifdef ALLEGRO_CFG_SHADER_HLSL
 
 static const char *default_hlsl_pixel_source =
    "bool " ALLEGRO_SHADER_VAR_USE_TEX ";\n"
